requirement.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
indic-nlp-library==0.6
joblib==0.16.0
Morfessor==2.0.6
murmurhash==1.0.2
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.91
six==1.15.0
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
torch==1.6.0
torchtext==0.6.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
wasabi==0.7.1
wincertstore==0.2
zipp==3.1.0
pytesseract==0.3.7
gtts==2.2.2
pyglet==1.5.16
